Algorithms

Equalization of Brightness

The Equalization of Brightness algorithm normalizes the amplitude of an image without change of phases and therefore of structure.

In general the algorithm can be applied to any image image which resembles a 2D array through Eob(image). In this case all default values are created and can be reused. For more complex and sophisticated situations wiht mnore control, it is recommended to use the EobAlgorithm command, where the filterbanks and the regularitzation parameter can be stated explicitely. To help with the creation of own steerable Riesz filterbanks, the function EobInitialization can be used.

RieszFrames.Eob — Function

Eob(input[, type=Float64])

Equalization of Brighness (Eob) with default values and new Initializationof the steerable filterbanks.

Arguments

input: this is the input grayscale image of any size. It will be converted into an according floating point structure during runtime

type: The type specifies the floating point type of the filterbanks (default: Float64).

Example

julia> Eob(rand(Float32,16,16));

RieszFrames.EobInitialization — Function

EobInitialization(inputsize[, maxscale=8, type=Float64, framefunction=Held])

Calculate the Filters and all preallocations for the EoB Algorithm for images of size in input

This function is especially usefulf for the generation of a new filterbank which is then used for batch processing of the Eob Algorithm. All important parameters can be set individually. The function wraps the calls to GenerateFrameFilters2D The functions of Generation Functions give more explanation about the parameters

Example

julia> testImage = rand(Float64,16,16);
julia> EobInitialize(size(testImage));

RieszFrames.EobAlgorithm — Function

EobAlgorithm(input[, regularization=0.001, framefilterbank=framebank[], rieszfilterbank=rieszbank[], type=Float64])

Calculate the EqualizationOfBrightness Algorithm with given values. It uses the default Values from the Initialization

This function is for more sophisticated applications, where the algorithm is batch-processed and the filterbanks are reused every time. This improves the performance because they can be reused. additionally the regularization parameter can be set individually. The applicaiton of the filterbanks is multithreaded

Arguments

input: input image. Must be a 2D Array of a floating point type

regularization: this parameter controls the regularization of the algorithm. 0 is a possible input for no regularization (default: 0.001)

framefilterbank: a valid frame filterbank array for the input image (default: global initialized framefilterbank)

rieszfilterbank: a valid riesz filterbank array for the input image (default: global initialized rieszfilterbank)

type: The type specifies the floating point type of the result (default: Float64).

Example

julia> result = EobAlgorithm(testImage);

multiresolution decomposition

This algorithm will follow in a future release

multireolution orientaiton estimation

This algorithm will follow in a future release