Some Frame Functions

The following frame functions are possible to use as the framefunction argument. The theory behind the construction is documented here

RieszFrames.Held — Function

Held(x,[,m])

generates the Held-Wavelet with the smoothness m

The Wavelet is constructed according to the paper S. Held, M. Storath, P. Massopust and B. Forster, "Steerable Wavelet Frames Based on the Riesz Transform" doi: 10.1109/TIP.2009.2036713

RieszFrames.Shannon — Function

Shannon(x)

generates theShannon Wavelet Frame

RieszFrames.Papadakis — Function

Papadakis(x)

generates the Papadakis Wavelet Frame

RieszFrames.Simoncelli — Function

Simoncelli(x)

generates the Simoncelli Wavelet Frame