Generation Functions

These functions help to generate data to get more control of the algorithm

RieszFrames.GenerateFrameFilters2D — Function

GenerateRieszFilters(filtersize[, maxscale, framefunction, type])

Generate the Frame filterbank for the according filtersize and choosen Motherwavelet

Arguments

maxscale: blabla

framefunction: bla

type: some info about type

RieszFrames.GenerateRieszFilters2D — Function

GenerateRieszFilters2D(filtersize[, maxrieszorder, type])

Generate the Riesz filterbank for the according filtersize and depth of Riesz

RieszFrames.GenerateRadius2D — Function

GenerateRadius2D(filtersize[, type])

Generates the Radius for a given 2D filtersize

RieszFrames.GenerateRadius1D — Function

GenerateRadius1D(filtersize[, type])

Generates the Radius for a given 1D filtersize