Generation Functions

filterbank generations

These functions help to generate data to get more control of the algorithm

RieszFrames.GenerateFrameFilters2D — Function

GenerateFrameFilters2D(filtersize[, maxscale, framefunction, type])

Generate the Frame filterbank for the according filtersize and choosen Motherwavelet

Arguments

maxscale: The amount of downsampling steps the filter is generated for (default: 8).

framefunction: This is the function of the motherwavelet (default: Held).

type: The type specifies the floating point type of the filterbanks (default: Float64).

Example

julia> filterbank = GenerateFrameFilters2D((64, 64))
julia> filterbank = GenerateFrameFilters2D((32, 32), 3, Held, Float32)

RieszFrames.GenerateRieszFilters2D — Function

GenerateRieszFilters2D(filtersize[, maxrieszorder, type])

Generate the Riesz filterbank for the according filtersize and depth of Riesz

Higher order Riesz transforms can be achieved by raising the umber of maxrieszorder. The according filterbank is of a pyramid scheme. The filtersize is a tuple

Example

julia> filterbank = GenerateFrameFilters2D((128, 64))
julia> filterbank = GenerateFrameFilters2D((32, 32), 3, Papadakis, Float64)

utility functions

RieszFrames.GenerateRadius1D — Function

GenerateRadius1D(filtersize[, type])

Generates the Radius for a given 1D filtersize

RieszFrames.GenerateRadius2D — Function

GenerateRadius2D(filtersize[, type])

Generates the Radius for a given 2D filtersize